+++ app/controllers/registrations_controller.rb
@@ -3,7 +3,6 @@
 # Adds a recaptcha on registration.
 class RegistrationsController < Devise::RegistrationsController
   prepend_before_action :check_captcha, only: [:create]
-  append_after_action :generate_welcome_email, only: [:create]
 
   layout "layouts/full_page"
 
@@ -14,9 +13,5 @@ class RegistrationsController < Devise::RegistrationsController
     self.resource = resource_class.new sign_up_params
     resource.errors.add(:recaptcha, "reCAPTCHA verification failed.")
     respond_with_navigational(resource) { render :new }
-  end
-
-  def generate_welcome_email
-    current_user&.send_welcome_email_in_background!
   end
 end