+++ test/controllers/admin_controller_test.rb
@@ -48,6 +48,15 @@ class AdminControllerTest < ActionDispatch::IntegrationTest
     assert_redirected_to admin_spam_inbox_path
   end
 
+  test "should destroy spammer as admin" do
+    login(@admin)
+    assert_difference("User.where(spammer: true).count") do
+      post admin_destroy_spammer_url(users(:shadow_banned), format: "js")
+    end
+    assert_template "destroy_spammer"
+    assert_response :success
+  end
+
   test "should unspamban user as admin" do
     login(@admin)
     assert_difference("User.where(spammer: nil).count", -1) do